+++ CHANGELOG.md
@@ -6,3 +6,5 @@ Practice mode:
     - These rounds are un-judged, and do not contribute to the remainder of the test.
     - Roughly 4 correct answers in a row with an art less than ``right_count_art_less_than`` will continue on to self-paced mode. This is the only path to continue to self-paced mode.
     - If more than 20 screens pass without the above-mentioned condition, the test will exit unsuccessfully with error code 4.
+
+- Merges self-paced into practice-test